🐛 fix: PDF 뷰어 배포 호환성 개선
- viewer 진입 경로가 .js 모듈 자산만 요청하도록 변경

- PDF.js .mjs 자산을 .js 별칭으로 동기화

- 모듈 로드 실패 시 PDF 직접 열기 fallback 제공
